fix(watcher): improve watching efficiency
This fix significantly improves watching performance. It tells chokidar to ignore all paths that do not match any of the watched patterns. Before we were only ignoring paths that match some of the excludes patterns. That means that even files that didn't match any of the watched patterns were still watched by chokidar and ignored on the level of `fileList`.

Note, this fix requires https://github.com/paulmillr/chokidar/pull/61.

Closes #616